shift prefill call to worker
