+++ blocks/form/util.js
@@ -1,5 +1,6 @@
 // create a string containing head tags from h1 to h5
 import { defaultErrorMessages } from './constant.js';
+import { externalize } from './rules/functions.js';
 
 const headings = Array.from({ length: 6 }, (_, i) => `<h${i + 1}>`).join('');
 const allowedTags = `${headings}<a><b><p><i><em><strong><ul><li><ol>`;
@@ -424,5 +425,18 @@ export function createDropdownUsingEnum(fd, wrapper) {
 
   if (ph && optionSelected === false) {
     ph.setAttribute('selected', '');
+  }
+}
+
+export async function fetchData(id, search = '') {
+  try {
+    const url = externalize(`/adobe/forms/af/data/${id}${search}`);
+    const response = await fetch(url);
+    const json = await response.json();
+    const { data: prefillData } = json;
+    const { data: { afData: { afBoundData: { data = {} } = {} } = {} } = {} } = json;
+    return Object.keys(data).length > 0 ? data : (prefillData || json);
+  } catch (ex) {
+    return null;
   }
 }
